Add instrumentation
Hook our React app up to Prometheus.

Rant time.

It is MIND-BOGGLING that there are no tutorials on how to do this. I
spent two days hacking the entire Earth to find a way of adding the
`/metrics` route to this app, including:

- Adding it as a component (HA HA; React is for SPAs, so that's a no-dice),
- Manually crafting a `<Route>` object (See the first bullet),
- Ejecting the application's configuration using `npm run eject` and
  adding it to the webpack config directly (to only learn that webpack
  just compiles everything into static assets and bundles them, even
  though it is _also_ somehow responsible for starting a web
  server...somewhere), and
- Unejecting (it's irreversible...if you don't know Git) the app,
  adding a `serverProxy.js` and having the proxy create the metrics via
  interception (which doesn't work because the proxy keeps routing to
  itself, as you would expect).

It turns out that all you need to do is compile the app with `npm
build`, then create an Express config at the root of the project, point
it at `index.html` and run `node server.js` (assuming the file is called
`server.js`).

It was only after adding the `/metrics` route there did I learn that
`express-prom-bundle` exists which creates the exporter for you
extremely easily.

You know a blog post is coming to address this.